// src/app/Core/Models/ReviewModel.ts
    export interface ReviewModel {
        movieId : number;
        rating:  number;
        reviewText: string;
    }


// Generated by https://quicktype.io

export interface UserReviewModel {
    movieId:    number;
    userId:     number;
    rating:     number;
    reviewText: string;
    posterURL:  string;
    title:      string;
}